+++ packages/gasket-plugin-engine/lib/flatten.js
@@ -1,3 +1,0 @@
-module.exports = function flatten(arrayOfArrays) {
-  return arrayOfArrays.reduce((result, array) => result.concat(array), []);
-};