+++ packages/gasket-plugin-engine/test/dependencies.test.js
@@ -1,29 +1,27 @@
-describe('The PluginEngine constructor', () => {
+describe('Plugin dependencies', () => {
   it('throws an Error if a required dependency of a plugin is missing', () => {
     const pluginA = {
+      name: 'pluginA',
       dependencies: ['testb', 'testc'],
       hooks: {}
     };
 
     const pluginB = {
+      name: 'pluginB',
       hooks: {}
     };
 
-    const pluginC = {
-      hooks: {}
-    };
-
-    jest
-      .doMock('@gasket/testa-plugin', () => pluginA, { virtual: true })
-      .doMock('@gasket/testb-plugin', () => pluginB, { virtual: true })
-      .doMock('@gasket/testc-plugin', () => pluginC, { virtual: true });
-
-    const PluginEngine = require('..');
-    const Resolver = require('../lib/resolver');
-    jest.spyOn(Resolver.prototype, 'tryResolve').mockImplementation(arg => {
-      return `${process.cwd()}/node_modules/${arg}`;
+    const { Loader } = require('@gasket/resolve');
+    jest.spyOn(Loader.prototype, 'loadConfigured').mockImplementation(() => {
+      return {
+        plugins: [
+          { module: pluginA },
+          { module: pluginB }
+        ]
+      };
     });
 
-    expect(() => new PluginEngine({ plugins: { add: ['testa', 'testb'] } })).toThrow(Error);
+    const PluginEngine = require('..');
+    expect(() => new PluginEngine({})).toThrow(Error);
   });
 });